Use nbsp as placeholder
Originally, <li>&shy;</li> is used as a placeholder (in the search
and related notes result list), and we use clientHeight to measure
the space taken by a list item, thereby calculating the number of
results we can fit into the list.

However, it turned out that Firefox will treat <li>&shy;</li> as
empty and thus collapse it to have height 0 (unlike Chromium).
Therefore we now use <li>&nbsp;</li> instead.