Use /static/index.html as PWA index
The service worker can only control urls under its scope,
which is restricted to /static/ since worker.js is served
from /static/worker.js. Thus / is not under this scope.

To work around this, we instead specify /static/index.html
as the start_url of the PWA. The service worker will respond
with whatever data it cached for / instead.