Fix autosave .innerText is undefined
Originally, the code reads "[...]?.innerText", which worked fine,
except we now want to trim the innerText. This was thus changed to
"[...]?.innerText.trim()", which will fail if "[...]" is undefined,
and thus "[...]?.innerText" is undefined, and has no attribute trim.

Changing that to "[...]?.innerText?.trim()" will not help, because
while the undefinedness propagates fine upto trim, undefined is not
a function and cannot be called.

Thus we have to introduce a helper function trim(str), that returns
undefined if str?.trim is undefined, and calls str.trim() otherwise.